+++ src/components/viewer/Viewer.vue
@@ -573,7 +573,7 @@ export default defineComponent({
         preventDragOffset: 40,
       });
 
-      // Live photo support
+      // Live Photo support
       new PsLivePhoto(<any>this.photoswipe, {});
 
       // Image support
@@ -866,10 +866,10 @@ export default defineComponent({
       // Only for JPEG for now
       if (!this.canEdit) return;
 
-      // Prevent editing live photos
+      // Prevent editing Live Photos
       if (this.currentPhoto.liveid) {
         alert(
-          this.t("memories", "Editing is currently disabled for live photos")
+          this.t("memories", "Editing is currently disabled for Live Photos")
         );
         return;
       }