+++ CHANGELOG.md
@@ -19,7 +19,7 @@ You may need to clear browser cache to use location search.
 - **Feature**: Better configurability and feature detection for go-vod ([#450](https://github.com/pulsejet/memories/issues/450))
 - **Feature**: Configurable folder/album sorting order ([#371](https://github.com/pulsejet/memories/issues/371))
 - **Feature**: Configurable album list sorting order ([#377](https://github.com/pulsejet/memories/issues/377))
-- **Feature**: Allow archiving photos throw folder view ([#350](https://github.com/pulsejet/memories/issues/350))
+- **Feature**: Allow archiving photos through folder view ([#350](https://github.com/pulsejet/memories/issues/350))
 - **Feature**: Add search bar to face cluster merge dialog ([#177](https://github.com/pulsejet/memories/issues/177))
 - **Bugfix**: Sidebar now shows metadata on albums and public shares ([#320](https://github.com/pulsejet/memories/issues/320)).
 - Other fixes and features ([milestone](https://github.com/pulsejet/memories/milestone/9?closed=1))